Factor out keys computation into separate module
